D8-207 (#41)
* D8-207

Grants the ability to upload images to the server. The component may be added wherever it is necessary. This includes placing it in admin only areas wherever needed.

* D8-208

Saves images on the server via an admin only form. Checks to ensure if Upload is in use, it is in use by an admin.

* D8-207

Forces the uploads folder to be made so errors won't occur

* D8-207

Minor Visual update

* D8-207

Minor Bug fixes and visual alterations